feat(csc-it-services-2025): simplify example
Do not treat `gendata.C` and `fitdata.C` file names as "Snakemake
inputs" from the scientific point of view. Only the "number of events"
is a scientific input to the workflow. The analysis file names are not
likely to change and can therefore be hardcoded in the Snakemake
workflow specification.